moved all tests to seperate folder
